chang type of (mm) to deciaml

+++ sql_scripts/create_iso_container_size_code_1.sql
@@ -6,7 +6,7 @@ GO
 
 CREATE TABLE ISOSizeCode1
 (
-    [Container length (mm)] int,
+    [Container length (mm)] decimal,
     [Container length (ft)] decimal,
     [Container length (in)] decimal,
     [Code character]        nvarchar(255)
